requirements.txt
cmake
cython
h5py
tqdm
seaborn
matplotlib
Pillow
Flask
SQLAlchemy
beautifulsoup4
jupyter
jupyterlab
numpy
pandas
scipy
scikit-learn
nltk
spacy
PyGithub
jsonlines
pyyaml
pydriller
GitPython